+++ src/components/stable/gux-calendar/gux-calendar.types.ts
@@ -9,3 +9,5 @@ export interface IDateElement {
   disabled: boolean;
   hidden: boolean;
 }
+
+export type GuxCalendarDayOfWeek = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;